decentralize pretty-format types (#7972)
